+++ core.php
@@ -54,7 +54,7 @@ function GASSAutoload($name) {
 		$classFound = false;
 		foreach ($includePaths as $includePath) {
 			$proposedPath = $includePath . DIRECTORY_SEPARATOR . $filePath . '.php';
-			if (file_exists($proposedPath) && is_readable($proposedPath)) {
+			if (@file_exists($proposedPath) && @is_readable($proposedPath)) {
 				$fileFound = true;
 				require_once $proposedPath;
 				if (class_exists($name) || interface_exists($name)) {